Fungus is amazing and I love it, great progress
Should have used this from the start but its whatever, probably only
saved me half an hour or so. At least we tried, but in the end Fungus
was always superior.

Not gonna come back and clean the code, instead just gonna do it
properly next time which works the same anyway. If I do come back, I
would like to complete the Order UI then I'll be truly satisfied.