We are actually really tired though
half done with the 2d Platformer tutorial, should have complete game
controls by then and the UI stat bars.

After that we need to implement the three system on our own.

The enter building should be the simplest one
because we just fade inbetween some house scenes. I wouldn't prefer it
but, theres only gonna get at least 5 maximum so it won't be too much of
a big deal and its only in the first Act.

Act 2's bypass script shouldn't be too much more complicated either. We
reuse the same IEnemyState and we make a new variant called Bypass which
only has patrol and alert.

Patrol is just normal patrol but from a point to a point instead of a
timer, and Alert will just check if the player is in the sight of the
main road which will then make the car shoot hundreds of bullets to
instantly kill you.

Last one is the base defence script. Might be a little complicated idk,
we need to send waves of enemies to attack the base instead of us. Base
will have hp like player's and if it depletes to 0 a gameover cutscene
will commence which will happen again at Act4 regardless.

In Act3 we will put some base defender AI's, just fire when they come in
sight of them, should be really easy in theory.

Last but not least are the polishing coding issues, most likely won't
have enough time to polish the empty animations to fill in so the
functions are more important.

Need to add the ingame-chat function with different UI, the script will
just use order so also relatively simple.

That should be 100% coding done that we can afford to do.

I think we will also have to record the sound effects first before we do
the art since it need proper timing, so from 12am onwards we will try to
get all the sounds then just draw from that point onwards until the
final deadline arrives