1 week was not enough XD
audio is priority when we wake up for the final day. At this point the
ending is the priority so we will do act 3&4 to the end before anything
else so we are safe.

If we do have enough extra time, we will do Act2 and as much polishling
as possible.